// fc-monolito-checkout/src/modules/client-adm/usecase/add-client/add-client.dto.ts
export interface AddClientInputDto {
  id?: string;
  name: string;
  document: string;
  email: string;
  street: string;
  number: string;
  complement: string;
  city: string;
  state: string;
  zipCode: string;
  createdAt?: Date;
  updatedAt?: Date;
}

export interface AddClientOutputDto {
  id: string;
  name: string;
  document: string;
  email: string;
  street: string;
  number: string;
  complement: string;
  city: string;
  state: string;
  zipCode: string;
  createdAt?: Date;
  updatedAt?: Date;
}